fix: refactor swc and combine both swc_readers into a combined method
